moves crucial part of introduceTopLevel to Run
This is done to make it easy for other parts of the compiler to use
the (quite involved) logic of introduceTopLevel.
