+++ src/compiler/scala/reflect/macros/runtime/Synthetics.scala
@@ -5,7 +5,6 @@
 package scala.reflect.macros
 package runtime
 
-import java.util.UUID._
 import scala.reflect.internal.Flags._
 import scala.reflect.internal.util.BatchSourceFile
 import scala.reflect.io.VirtualFile
@@ -42,11 +41,6 @@ trait Synthetics {
     else EmptyTree
   }
 
-  // TODO: provide a way to specify a pretty name for debugging purposes
-  private def randomFileName() = (
-    "macroSynthetic-" + randomUUID().toString.replace("-", "") + ".scala"
-  )
-
   def introduceTopLevel[T: PackageSpec](packagePrototype: T, definition: universe.ImplDef): RefTree =
     introduceTopLevel(packagePrototype, List(definition)).head
 
@@ -55,18 +49,7 @@ trait Synthetics {
 
   private def introduceTopLevel[T: PackageSpec](packagePrototype: T, definitions: List[universe.ImplDef]): List[RefTree] = {
     val code @ PackageDef(pid, _) = implicitly[PackageSpec[T]].mkPackageDef(packagePrototype, definitions)
-    val syntheticFileName = randomFileName()
-    // compatibility with SBT
-    // on the one hand, we need to specify some jfile here, otherwise sbt crashes with an NPE (SI-6870)
-    // on the other hand, we can't specify the obvious enclosingUnit, because then sbt somehow fails to run tests using type macros
-    // okay, now let's specify a guaranteedly non-existent file in an existing directory (so that we don't run into permission problems)
-    val relatedJfile = enclosingUnit.source.file.file
-    val fakeJfile = if (relatedJfile != null) new java.io.File(relatedJfile.getParent, syntheticFileName) else null
-    val virtualFile = new VirtualFile(syntheticFileName) { override def file = fakeJfile }
-    val sourceFile = new BatchSourceFile(virtualFile, code.toString)
-    val unit = new CompilationUnit(sourceFile)
-    unit.body = code
-    universe.currentRun.compileLate(unit)
+    universe.currentRun.compileLate(code)
     definitions map (definition => Select(pid, definition.name))
   }
